+++ lib/glimmer/dsl/swt/cursor_expression.rb
@@ -1,5 +1,4 @@
 require 'glimmer/dsl/expression'
-require 'glimmer/dsl/top_level_expression'
 require 'glimmer/swt/cursor_proxy'
 
 module Glimmer
@@ -8,11 +7,9 @@ module Glimmer
       # cursor expression
       # Note: Cannot be a static expression because it clashes with cursor property expression
       class CursorExpression < Expression
-        include TopLevelExpression
-  
         def can_interpret?(parent, keyword, *args, &block)
           keyword.to_s == 'cursor' and
-          (parent.nil? or !parent.respond_to?('cursor')) and
+            (parent.nil? or !parent.respond_to?('cursor')) and
             args.size == 1 and
             (args.first.is_a?(Integer) or textual?(args.first))
         end